dobj matching

+++ include/uv_model.h
@@ -3,7 +3,7 @@
 
 #include <uv_matrix.h>
 
-void uvModelGetProps(u16, s32, void*, s32);
+void uvModelGetProps(s32, s32, void*, s32);
 void uvModelGetPosm(s32 model, s32 part, Mtx4F*);
 
 #endif // UV_MODEL_H
